% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.droplasso.R
\name{coef.droplasso}
\alias{coef.droplasso}
\title{Coefficients of a model of class droplasso}
\usage{
\method{coef}{droplasso}(object, s = NULL, ...)
}
\arguments{
\item{object}{Object of class \code{droplasso}}

\item{s}{Value(s) of the penalty parameter lambda at which coefficients are 
required. Can be either omitted, or a vector of numeric values. If omitted,
default is the entire sequence used to create the model.}

\item{...}{other arguments.}
}
\value{
A matrix of coefficients, each column corresonds to a regularization 
  parameter.
}
\description{
This function extracts the matrix of coefficients for the model(s) trained by
the \code{droplasso} function corresponding to regularization parameters of 
interest.
}
\references{
Adapted from the same function in 
  \href{https://cran.r-project.org/web/packages/Coxnet/index.html}{Coxnet} 
  package.
}
\seealso{
\code{droplasso}, \code{predict} methods.
}
