% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.droplasso.R
\name{plot.droplasso}
\alias{plot.droplasso}
\title{Plot coefficients from a "droplasso" object}
\usage{
\method{plot}{droplasso}(m, sign.lambda = 1, ...)
}
\arguments{
\item{m}{Fitted model of class "survenet"}
}
\value{
None
}
\description{
Plot coefficients from a "droplasso" object
}
\examples{
#create data:
nobs = 100
nvars = 5
x = matrix(rnorm(nobs*nvars),nrow=nobs)
b = c(1,1,0,0,0)
p = 1/(1+exp(-x\%*\%b))
y = p>0.5
# Fit a dropout lasso model
m <- droplasso(x, y, family="binomial", nlambda=50, decay=5)
# Plot is
plot(m)
}
