% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.droplasso.R
\name{predict.droplasso}
\alias{predict.droplasso}
\title{Make predictions from a droplasso object.}
\usage{
\method{predict}{droplasso}(object, newx, s = NULL, type = c("link",
  "coefficients"), ...)
}
\arguments{
\item{object}{Fitted "\code{droplasso}" model object}

\item{newx}{Matrix of new values for \code{x} at which predictions are to be 
made. This argument is not used for \code{type="coefficients"}.}

\item{s}{Value(s) of the penalty parameter lambda at which predictions are 
required. Can be either omitted, or a vector of numeric values. If omitted
(or equal to \code{NULL}), default is the entire sequence used to create
the model.}

\item{type}{Type of prediction required. Type \code{"link"} (default) gives 
the linear predictor. Type \code{"coefficients"} computes the coefficients 
at the requested values of \code{s}. If \code{newx} is not provided, then 
\code{type="coefficients"} automatically.}

\item{...}{Other parameters.}
}
\value{
If \code{type="link"}, returns a matrix of prediction scores, one row
  per sample of \code{newx}, one column per value of lambda. If 
  \code{type="coefficients"}, returns a matrix of weights of the model, one 
  row per feature, one column per value of lambda.
}
\description{
Similar to other \code{predict} methods, this function predicts fitted values
or coefficients from a fitted "\code{droplasso}" object
}
\examples{
#create data:
nobs = 100
nvars = 5
x = matrix(rnorm(nobs*nvars),nrow=nobs)
b = c(1,1,0,0,0)
p = 1/(1+exp(-x\%*\%b))
y = p>0.5
# Fit a dropout lasso model
m <- droplasso(x, y, family="binomial", nlambda=50, decay=5)
# Plot is
plot(m)
# Plot the weights for lambda=0.01
plot(predict(m,s=0.01), xlab="Features", ylab="Weight", main="Lambda=0.01")
# Plot prediction score
plot(predict(m, x, s=0.01), xlab="Sample", ylab="Score", main="Lambda=0.01")
}
\seealso{
\code{droplasso}, \code{coef} methods.
}
